# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayRaider API",
    "description": "API for Stellar network analytics, anchor monitoring, and payment corridor insights",
    "contact": {
      "name": "PayRaider Team",
      "email": "support@payraider.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.payraider.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "GET /webhooks - List webhooks for authenticated user",
        "operationId": "list_webhooks",
        "responses": {
          "200": {
            "description": "List of webhooks"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "POST /webhooks - Register a new webhook",
        "operationId": "register_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Webhook registered successfully"
          },
          "400": {
            "description": "Invalid webhook URL or missing event types"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/webhooks/{id}": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "GET /webhooks/:id - Get a single webhook by ID",
        "operationId": "get_webhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Webhook ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner"
          },
          "404": {
            "description": "Webhook not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Webhooks"
        ],
        "summary": "DELETE /webhooks/:id - Delete/deactivate webhook",
        "operationId": "delete_webhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Webhook ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Webhook not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/webhooks/{id}/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "POST /webhooks/:id/test - Queue a test event for delivery",
        "operationId": "test_webhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Webhook ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test event queued"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner"
          },
          "404": {
            "description": "Webhook not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWebhookRequest": {
        "type": "object",
        "description": "Webhook creation request",
        "required": [
          "url",
          "event_types"
        ],
        "properties": {
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filters": {},
          "url": {
            "type": "string"
          }
        }
      },
...
      "WebhookResponse": {
        "type": "object",
        "description": "Webhook creation response",
        "required": [
          "id",
          "url",
          "event_types",
          "is_active",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filters": {},
          "id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}